[Code] move nodegit related tests from jest to mocha
