+++ package.json
@@ -371,7 +371,7 @@
     "mocha": "3.3.0",
     "murmurhash3js": "3.0.1",
     "mutation-observer": "^1.0.3",
-    "nock": "8.0.0",
+    "nock": "10.0.4",
     "node-sass": "^4.9.4",
     "normalize-path": "^3.0.0",
     "pixelmatch": "4.0.2",
@@ -381,7 +381,7 @@
     "proxyquire": "1.7.11",
     "regenerate": "^1.4.0",
     "simple-git": "1.37.0",
-    "sinon": "^5.0.7",
+    "sinon": "v7.2.2",
     "strip-ansi": "^3.0.1",
     "supertest": "^3.1.0",
     "supertest-as-promised": "^4.0.2",
