+++ x-pack/plugins/code/server/__tests__/lsp_indexer_test.ts
@@ -4,22 +4,20 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
+import assert from 'assert';
 import fs from 'fs';
 import Git, { CloneOptions } from 'nodegit';
 import path from 'path';
 import rimraf from 'rimraf';
 import sinon from 'sinon';
-
+import { LspIndexer } from '../indexer/lsp_indexer';
+import { RepositoryGitStatusReservedField } from '../indexer/schema';
 import { AnyObject, EsClient } from '../lib/esqueue';
 import { Log } from '../log';
 import { InstallManager } from '../lsp/install_manager';
 import { LspService } from '../lsp/lsp_service';
 import { ServerOptions } from '../server_options';
 import { ConsoleLoggerFactory } from '../utils/console_logger_factory';
-import { LspIndexer } from './lsp_indexer';
-import { RepositoryGitStatusReservedField } from './schema';
-
-jest.setTimeout(30000);
 
 const log: Log = (new ConsoleLoggerFactory().getLogger(['test']) as any) as Log;
 
@@ -148,124 +146,130 @@ function setupLsServiceSendRequestSpy(): sinon.SinonSpy {
     })
   );
 }
-
-beforeAll(async () => {
-  return new Promise(resolve => {
-    rimraf(serverOptions.repoPath, resolve);
+describe('lsp_indexer', () => {
+  // @ts-ignore
+  before(async () => {
+    return new Promise(resolve => {
+      rimraf(serverOptions.repoPath, resolve);
+    });
   });
-});
 
-beforeEach(async () => {
-  await prepareProject(
-    'https://github.com/Microsoft/TypeScript-Node-Starter.git',
-    path.join(serverOptions.repoPath, repoUri)
-  );
-});
+  beforeEach(async function() {
+    // @ts-ignore
+    this.timeout(200000);
+    return await prepareProject(
+      'https://github.com/Microsoft/TypeScript-Node-Starter.git',
+      path.join(serverOptions.repoPath, repoUri)
+    );
+  });
+  // @ts-ignore
+  after(() => {
+    return cleanWorkspace();
+  });
 
-afterAll(() => {
-  return cleanWorkspace();
-});
+  afterEach(() => {
+    sinon.restore();
+  });
 
-afterEach(() => {
-  sinon.restore();
-});
+  it('Normal LSP index process.', async () => {
+    // Setup the esClient spies
+    const {
+      getSpy,
+      existsAliasSpy,
+      createSpy,
+      putAliasSpy,
+      deleteByQuerySpy,
+      bulkSpy,
+    } = setupEsClientSpy();
 
-test('Normal LSP index process.', async () => {
-  // Setup the esClient spies
-  const {
-    getSpy,
-    existsAliasSpy,
-    createSpy,
-    putAliasSpy,
-    deleteByQuerySpy,
-    bulkSpy,
-  } = setupEsClientSpy();
+    const lspservice = new LspService(
+      '127.0.0.1',
+      serverOptions,
+      esClient as EsClient,
+      {} as InstallManager,
+      new ConsoleLoggerFactory()
+    );
 
-  const lspservice = new LspService(
-    '127.0.0.1',
-    serverOptions,
-    esClient as EsClient,
-    {} as InstallManager,
-    new ConsoleLoggerFactory()
-  );
+    lspservice.sendRequest = setupLsServiceSendRequestSpy();
 
-  lspservice.sendRequest = setupLsServiceSendRequestSpy();
+    const indexer = new LspIndexer(
+      'github.com/Microsoft/TypeScript-Node-Starter',
+      'master',
+      lspservice,
+      serverOptions,
+      esClient as EsClient,
+      log as Log
+    );
+    await indexer.start();
 
-  const indexer = new LspIndexer(
-    'github.com/Microsoft/TypeScript-Node-Starter',
-    'master',
-    lspservice,
-    serverOptions,
-    esClient as EsClient,
-    log as Log
-  );
-  await indexer.start();
+    // Expect EsClient get called once to get the repo git status.
+    assert.ok(getSpy.calledOnce);
 
-  // Expect EsClient get called once to get the repo git status.
-  expect(getSpy.calledOnce).toBeTruthy();
+    // Expect EsClient deleteByQuery called 3 times for repository cleaning before
+    // the index for document, symbol and reference, respectively.
+    assert.strictEqual(deleteByQuerySpy.callCount, 3);
 
-  // Expect EsClient deleteByQuery called 3 times for repository cleaning before
-  // the index for document, symbol and reference, respectively.
-  expect(deleteByQuerySpy.callCount).toEqual(3);
+    // Ditto for index and alias creation
+    assert.strictEqual(existsAliasSpy.callCount, 3);
+    assert.strictEqual(createSpy.callCount, 3);
+    assert.strictEqual(putAliasSpy.callCount, 3);
 
-  // Ditto for index and alias creation
-  expect(existsAliasSpy.callCount).toEqual(3);
-  expect(createSpy.callCount).toEqual(3);
-  expect(putAliasSpy.callCount).toEqual(3);
+    // There are 22 files in the repo. 1 file + 1 symbol + 1 reference = 3 objects to
+    // index for each file. Total doc indexed should be 3 * 22 = 66, which can be
+    // fitted into a single batch index.
+    assert.ok(bulkSpy.calledOnce);
+    assert.strictEqual(bulkSpy.getCall(0).args[0].body.length, 66 * 2);
+    // @ts-ignore
+  }).timeout(20000);
 
-  // There are 22 files in the repo. 1 file + 1 symbol + 1 reference = 3 objects to
-  // index for each file. Total doc indexed should be 3 * 22 = 66, which can be
-  // fitted into a single batch index.
-  expect(bulkSpy.calledOnce).toBeTruthy();
-  expect(bulkSpy.getCall(0).args[0].body.length).toEqual(66 * 2);
-});
+  it('Cancel LSP index process.', async () => {
+    // Setup the esClient spies
+    const {
+      getSpy,
+      existsAliasSpy,
+      createSpy,
+      putAliasSpy,
+      deleteByQuerySpy,
+      bulkSpy,
+    } = setupEsClientSpy();
 
-test('Cancel LSP index process.', async () => {
-  // Setup the esClient spies
-  const {
-    getSpy,
-    existsAliasSpy,
-    createSpy,
-    putAliasSpy,
-    deleteByQuerySpy,
-    bulkSpy,
-  } = setupEsClientSpy();
+    const lspservice = new LspService(
+      '127.0.0.1',
+      serverOptions,
+      esClient as EsClient,
+      {} as InstallManager,
+      new ConsoleLoggerFactory()
+    );
 
-  const lspservice = new LspService(
-    '127.0.0.1',
-    serverOptions,
-    esClient as EsClient,
-    {} as InstallManager,
-    new ConsoleLoggerFactory()
-  );
+    lspservice.sendRequest = setupLsServiceSendRequestSpy();
 
-  lspservice.sendRequest = setupLsServiceSendRequestSpy();
+    const indexer = new LspIndexer(
+      'github.com/Microsoft/TypeScript-Node-Starter',
+      'master',
+      lspservice,
+      serverOptions,
+      esClient as EsClient,
+      log as Log
+    );
+    // Cancel the indexer before start.
+    indexer.cancel();
+    await indexer.start();
 
-  const indexer = new LspIndexer(
-    'github.com/Microsoft/TypeScript-Node-Starter',
-    'master',
-    lspservice,
-    serverOptions,
-    esClient as EsClient,
-    log as Log
-  );
-  // Cancel the indexer before start.
-  indexer.cancel();
-  await indexer.start();
+    // Expect EsClient get called once to get the repo git status.
+    assert.ok(getSpy.calledOnce);
 
-  // Expect EsClient get called once to get the repo git status.
-  expect(getSpy.calledOnce).toBeTruthy();
+    // Expect EsClient deleteByQuery called 3 times for repository cleaning before
+    // the index for document, symbol and reference, respectively.
+    assert.strictEqual(deleteByQuerySpy.callCount, 3);
 
-  // Expect EsClient deleteByQuery called 3 times for repository cleaning before
-  // the index for document, symbol and reference, respectively.
-  expect(deleteByQuerySpy.callCount).toEqual(3);
+    // Ditto for index and alias creation
+    assert.strictEqual(existsAliasSpy.callCount, 3);
+    assert.strictEqual(createSpy.callCount, 3);
+    assert.strictEqual(putAliasSpy.callCount, 3);
 
-  // Ditto for index and alias creation
-  expect(existsAliasSpy.callCount).toEqual(3);
-  expect(createSpy.callCount).toEqual(3);
-  expect(putAliasSpy.callCount).toEqual(3);
-
-  // Because the indexer is cancelled already in the begining. 0 doc should be
-  // indexed and thus bulk won't be called.
-  expect(bulkSpy.notCalled).toBeTruthy();
-});
+    // Because the indexer is cancelled already in the begining. 0 doc should be
+    // indexed and thus bulk won't be called.
+    assert.ok(bulkSpy.notCalled);
+  });
+  // @ts-ignore
+}).timeout(20000);
