+++ x-pack/tasks/test.js
@@ -29,9 +29,10 @@ export default (gulp, { mocha }) => {
   gulp.task('testserver', () => {
     const globs = [
       'common/**/__tests__/**/*.js',
+      'common/**/__tests__/**/*.ts',
       'server/**/__tests__/**/*.js',
+      'server/**/__tests__/**/*.ts',
     ].concat(forPluginServerTests());
-
     return gulp.src(globs, { read: false })
       .pipe(mocha(MOCHA_OPTIONS));
   });
